+++ src/ful/elements/temporals.mjs
@@ -45,7 +45,7 @@ class Instant extends ParsedElement {
 }
 
 class InputLocalDate extends Input {
-    static observed = ['value', 'readonly:presence', 'required:presence', 'min', 'max', 'step'];
+    static observed = ['value', 'readonly:presence', 'required:presence', 'placeholder', 'min', 'max', 'step'];
     _type() {
         return 'date';
     }
@@ -123,7 +123,7 @@ class InputLocalTime extends InputLocalDate {
 }
 
 class InputInstant extends Input {
-    static observed = ['value', 'readonly:presence', 'required:presence', 'min', 'max', 'step'];
+    static observed = ['value', 'readonly:presence', 'required:presence', 'placeholder', 'min', 'max', 'step'];
     _type() {
         return 'datetime-local';
     }